Adds in review corrections
